# compact: Implement Tree using Range