--- a/merkle/compact/tree_test.go
+++ b/merkle/compact/tree_test.go
@@ -32,17 +32,18 @@ import (
 // This check ensures that the compact Merkle tree contains the correct set of
 // nodes, i.e. the node on level i is present iff i-th bit of tree size is 1.
 func checkUnusedNodesInvariant(t *Tree) error {
-	size := t.size
+	size := t.Size()
+	nodes := t.getNodes()
 	sizeBits := bits.Len64(uint64(size))
-	if got, want := len(t.nodes), sizeBits; got != want {
+	if got, want := len(nodes), sizeBits; got != want {
 		return fmt.Errorf("nodes mismatch: have %v nodes, want %v", got, want)
 	}
 	for level := 0; level < sizeBits; level++ {
 		if size&1 == 1 {
-			if t.nodes[level] == nil {
+			if nodes[level] == nil {
 				return fmt.Errorf("missing node at level %d", level)
 			}
-		} else if t.nodes[level] != nil {
+		} else if nodes[level] != nil {
 			return fmt.Errorf("unexpected node at level %d", level)
 		}
 		size >>= 1