This change replaces internals of `compact.Tree` with the `compact.Range`.

The old `Hashes` method/format are now used only in tests, and will be removed
in a follow-up after tests refactoring.

#### Benchmarks
No significant changes.
```
Before:
BenchmarkAppendLeaf-12             	     500	   3214217 ns/op
BenchmarkAppendLeafNoVisitor-12    	     500	   3069281 ns/op

After:
BenchmarkAppendLeaf-12             	     500	   3118462 ns/op
BenchmarkAppendLeafNoVisitor-12    	     500	   3089277 ns/op
```

<!---
Describe your changes in detail here.
If this fixes an issue, please write "Fixes #123", substituting the issue number.
-->

### Checklist

<!---
Go over all the following points, and put an `x` in all the boxes that apply.
Feel free to not tick any boxes that don't apply to this PR (e.g. refactoring may not need a CHANGELOG update).
If you're unsure about any of these, don't hesitate to ask. We're here to help!
-->

- [ ] I have updated the [CHANGELOG](CHANGELOG.md).
  - Adjust the draft version number according to [semantic versioning](https://semver.org/) rules.
- [ ] I have updated [documentation](docs/) accordingly.